refactor: specify image location
