+++ src/assets/components/navigate/NavigatorCurve.tsx
