+++ src/components/Navigator.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PathName from 'constants/PathName';
-import { ReactComponent as AddShopIcon } from 'assets/AddShopIcon.svg';
-import MapIcon from 'assets/MapIcon';
-import NavigatorCurve from 'assets/NavigatorCurve';
-import ProfileIcon from 'assets/ProfileIcon';
+import { ReactComponent as AddShopIcon } from 'assets/components/navigate/AddShopIcon.svg';
+import MapIcon from 'assets/components/navigate/MapIcon';
+import NavigatorCurve from 'assets/components/navigate/NavigatorCurve';
+import ProfileIcon from 'assets/components/navigate/ProfileIcon';
 import Typography from 'components/Typography';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
