+++ src/components/map/UserMarker.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import UserImgSrc from 'assets/user.svg';
+import UserImgSrc from 'assets/components/navigate/user.svg';
 import { LocationType } from 'models/Location';
 import { MapMarker } from 'react-kakao-maps-sdk';
